+++ packages/ui/src/pages/sites/components/docs-site/layout.tsx
@@ -1,10 +1,10 @@
 import * as React from 'react';
 
-export type SitesLayoutProps = {
+export type DocsSiteLayoutProps = {
   children: React.ReactNode;
 };
 
-export function SitesLayout(props: SitesLayoutProps): JSX.Element {
+export function DocsSiteLayout(props: DocsSiteLayoutProps): JSX.Element {
   return (
     <div className="mx-auto max-w-[52rem] px-4 pb-28 sm:px-6 md:px-8 lg:max-w-6xl xl:px-12">
       {props.children}
